skills/adk_agent/requirements.txt
# ADK Skills Framework - Dynamic Loading Agent
# Core dependencies

google-adk>=0.1.0
litellm>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
psutil>=5.9.0

fastapi>=0.100.0
uvicorn>=0.20.0